+++ app/helpers/QRScan.js
